Do not use bundled rspec for test runs
This caused problems with Dispatch, which could not parse the output of
RSpec when it was run with bundle exec, and therefore always showed the
quickfix window irrespective of test suite success or failure, and could
not highlight paths of failing tests in the quickfix window.

I wonder if this is to do with 'bundle exec rspec' not being a
recognised 'compiler' for Dispatch. Further invesitgation needed.